gallery-backend/requirements.txt
astroid==1.6.5
backports.functools-lru-cache==1.5
certifi==2018.4.16
chardet==3.0.4
configparser==3.5.0
Django==1.11.13
django-allauth==0.36.0
django-environ==0.4.5
django-rest-auth==0.9.3
django-webpack-loader==0.6.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
enum34==1.1.6
futures==3.2.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==2.1.0
Pillow==5.1.0
PyJWT==1.6.4
pylint==1.9.2
python-openid==2.2.5
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
singledispatch==3.4.0.3
six==1.11.0
urllib3==1.23
wrapt==1.10.11
